Utilize single class for tabular data, separate out manager
