+++ mathesar_ui/src/stores/table-data/manager.ts
@@ -1,0 +1,42 @@
+import { TabularType } from '@mathesar/App.d';
+import type { DBObjectEntry, TableEntry, ViewEntry } from '@mathesar/App.d';
+import { TabularData } from './tabularData';
+
+const tableMap: Map<TableEntry['id'], TabularData> = new Map();
+const viewMap: Map<ViewEntry['id'], TabularData> = new Map();
+
+function get(type: TabularType, id: DBObjectEntry['id']): TabularData {
+  const tabularMap = type === TabularType.View ? viewMap : tableMap;
+  let entry = tabularMap.get(id);
+  if (!entry) {
+    entry = new TabularData(type, id);
+    tabularMap.set(id, entry);
+  }
+  return entry;
+}
+
+function remove(type: TabularType, id: DBObjectEntry['id']): void {
+  const tabularMap = type === TabularType.View ? viewMap : tableMap;
+  // destroy all objects in table
+  const entry = tabularMap.get(id);
+  if (entry) {
+    entry.destroy();
+    tabularMap.delete(id);
+  }
+}
+
+export function getTableContent(id: TableEntry['id']): TabularData {
+  return get(TabularType.Table, id);
+}
+
+export function getViewContent(id: ViewEntry['id']): TabularData {
+  return get(TabularType.View, id);
+}
+
+export function removeTableContent(id: TableEntry['id']): void {
+  remove(TabularType.Table, id);
+}
+
+export function removeViewContent(id: ViewEntry['id']): void {
+  remove(TabularType.View, id);
+}
